vectordb_setup/pyproject.toml
[tool.poetry]
name = "vectordb-setup"
version = "0.1.0"
description = "loading the document Q&A into elasticsearch"
authors = ["Mor Sondak <mor.sondak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pandas = "^2.2.2"
elasticsearch = "^8.15.0"
langchain-elasticsearch = "^0.2.2"
langchain-core = "^0.2.35"
langchain-huggingface = "^0.0.3"
langchain = "^0.2.15"
langchain-community = "^0.2.14"
faiss-cpu = "^1.8.0.post1"
tiktoken = "^0.7.0"
fastembed = "^0.3.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
tqdm = "^4.66.5"
jupyterlab = "^4.2.5"
ipywidgets = "^8.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
